Cave Dungeon (So far)
This is what I have so far for the Cave Dungeon. The fence posts will be swapped out with the shadow tiles when we figure that out, but for now, they are just walls making the mazes. The right room will be the first room unlocked and it will have 3 enemies (pots) and a chest (barrel). Then the left room will be unlocked and just have a maze and a chest. Then the final room will be the boss room.